bbram: move z_impl from public API header file to another file
This moves all the inline z_impl_* functions from the public
API header file into another file. These z_impl_* functions
are implementation details that do not need to live within
the public API header file. This de-clutters the public API
header file.
